requirements.txt
stripe>=11.5.0
python-dotenv>=1.0.1

